Extract the database importer to a new class.
