+++ community/dbms/src/main/java/org/neo4j/commandline/dbms/ImportCommand.java
@@ -19,7 +19,6 @@
  */
 package org.neo4j.commandline.dbms;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -33,14 +32,10 @@ import org.neo4j.commandline.admin.OutsideWorld;
 import org.neo4j.dbms.DatabaseManagementSystemSettings;
 import org.neo4j.graphdb.factory.GraphDatabaseSettings;
 import org.neo4j.helpers.Args;
-import org.neo4j.io.fs.FileUtils;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.impl.util.Converters;
 import org.neo4j.kernel.impl.util.Validators;
 import org.neo4j.server.configuration.ConfigLoader;
-
-import static org.neo4j.dbms.DatabaseManagementSystemSettings.database_path;
-import static org.neo4j.helpers.collection.MapUtil.stringMap;
 
 public class ImportCommand implements AdminCommand
 {
@@ -73,6 +68,7 @@ public class ImportCommand implements AdminCommand
 
     private final Path homeDir;
     private final Path configDir;
+    private final String[] allowedModes = {"database"};
 
     public ImportCommand( Path homeDir, Path configDir )
     {
@@ -83,17 +79,11 @@ public class ImportCommand implements AdminCommand
     @Override
     public void execute( String[] args ) throws IncorrectUsage, CommandFailed
     {
-        File from;
-        String database;
-
         Args parsedArgs = Args.parse( args );
         try
         {
             parsedArgs.interpretOption( "mode", Converters.<String>mandatory(), s -> s,
-                    Validators.inList( new String[]{"database"} ) );
-            database = parsedArgs.interpretOption( "database", Converters.<String>mandatory(), s -> s );
-            from = parsedArgs.interpretOption( "from", Converters.<File>mandatory(), Converters.toFile(),
-                    Validators.CONTAINS_EXISTING_DATABASE );
+                    Validators.inList( allowedModes ) );
         }
         catch ( IllegalArgumentException e )
         {
@@ -102,34 +92,23 @@ public class ImportCommand implements AdminCommand
 
         try
         {
-            Config config = loadNeo4jConfig( homeDir, configDir, database );
-            copyDatabase( from, config );
-            removeMessagesLog( config );
+            Config config = loadNeo4jConfig( homeDir, configDir );
+            DatabaseImporter databaseImporter = new DatabaseImporter( parsedArgs, config );
+            databaseImporter.doImport();
         }
         catch ( IOException e )
         {
             throw new RuntimeException( e );
         }
-
     }
 
-    private void copyDatabase( File from, Config config ) throws IOException
-    {
-        FileUtils.copyRecursively( from, config.get( database_path ) );
-    }
-
-    private void removeMessagesLog( Config config )
-    {
-        FileUtils.deleteFile( new File( config.get( database_path ), "messages.log" ) );
-    }
-
-    private static Config loadNeo4jConfig( Path homeDir, Path configDir, String databaseName )
+    private static Config loadNeo4jConfig( Path homeDir, Path configDir )
     {
         ConfigLoader configLoader = new ConfigLoader( settings() );
         Config config = configLoader.loadConfig( Optional.of( homeDir.toFile() ),
                 Optional.of( configDir.resolve( "neo4j.conf" ).toFile() ) );
 
-        return config.with( stringMap( DatabaseManagementSystemSettings.active_database.name(), databaseName ) );
+        return config;
     }
 
     private static List<Class<?>> settings()
